Add `maptype` parameter for a mapfile
This allows a user to specify a non-`file` map type for an `autofs::mapfile`.
No content will be created for the map, and (in non-Solaris automount systems),
the map will be referenced with the map type prefix.

Example:
```puppet:
autofs::mapfile{ 'auto.chroot':
  directory => '/chroot',
  maptype   => program,
}
```

This would cause the `auto.master` file to include (in non-Solaris systems):

    /chroot program:auto.chroot

and would _not_ create the `/etc/auto.chroot` file (leaving that up to the user).